fix: gloo deps
