+++ paddle/phi/core/distributed/CMakeLists.txt
@@ -12,16 +12,16 @@ endif()
 
 if(WITH_GLOO)
   cc_library(
-    gloo_comm_context
-    SRCS gloo_comm_context.cc
-    DEPS dense_tensor)
-
-  cc_library(
     gloo_utils
     SRCS gloo_utils.cc
-    DEPS gloo)
+    DEPS gloo dense_tensor enforce)
 
-  list(APPEND COMM_CONTEXT_MANAGER_DEPS gloo_comm_context gloo_store gloo_utils)
+  cc_library(
+    gloo_comm_context
+    SRCS gloo_comm_context.cc
+    DEPS gloo_utils)
+
+  list(APPEND COMM_CONTEXT_MANAGER_DEPS gloo_comm_context gloo_store)
 endif()
 
 cc_library(
