+++ paddle/phi/core/distributed/gloo_utils.h
@@ -88,10 +88,16 @@ namespace distributed {
 #endif
 
 template <typename T, typename P>
-void SetOutput(P* opts, phi::DenseTensor* tensor);
+void SetOutput(P* opts, phi::DenseTensor* tensor) {
+  opts->setOutput(reinterpret_cast<T*>(tensor->data()), tensor->numel());
+}
 
 template <typename T, typename P>
-void SetInput(P* opts, const phi::DenseTensor& tensor);
+void SetInput(P* opts, const phi::DenseTensor& tensor) {
+  // gloo only support mutable data input
+  opts->setInput(reinterpret_cast<T*>(const_cast<void*>(tensor.data())),
+                 tensor.numel());
+}
 
 // env preparation
 std::shared_ptr<gloo::transport::Device> CreateDeviceForInterface(
